Fixes/updates.
* Admin
  * validate is public method
  * logout does not redirect
  * hash_pass is deprecated
* Benchmark
  * memory methods return a double instead of a string
* Cache\File
  * delete returns boolean
* Config
  * set returns value
* Tea\Test
  * updated manpage
* Test
  * assertIsA renamed to assertType
  * assertNotA renamed to assertNotType
  * assertIsA and assertNotA tests for instance of class

+++ tfd/benchmark.php
@@ -29,12 +29,12 @@
 		
 		public function current_memory($decimal = 2){
 			// return memory usage in mb
-			return number_format(memory_get_usage() / 1024 / 1024, $decimal);
+			return (double)number_format(memory_get_usage() / 1024 / 1024, $decimal);
 		}
 		
 		public function peak_memory($decimal = 2){
 			// return memory in mb
-			return number_format(memory_get_peak_usage() / 1024 / 1024, $decimal);
+			return (double)number_format(memory_get_peak_usage() / 1024 / 1024, $decimal);
 		}
 	
 	}
